refactor(mitm-addon): split firewall auth boundaries
