+++ crates/runner/mitm-addon/src/auth.py
@@ -1,22 +1,16 @@
-"""Firewall auth header fetching, caching, and URL rewriting.
+"""Firewall auth flow orchestration and request mutation.
 
-Manages TTL-based caching of resolved auth headers and handles
-both standard header injection and auth.base URL rewriting.
+This module keeps mitmproxy HTTPFlow metadata, header/query injection,
+auth.base forwarding, local failure responses, and AWS SigV4 signing together.
+Cache state and platform API calls live in dedicated owner modules.
 """
 
-import asyncio
 import json
-import math
-import os
-import time
-import urllib.error
 import urllib.parse
-import urllib.request
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 from enum import Enum
-from typing import Protocol
 
-from mitmproxy import ctx, http
+from mitmproxy import http
 
 import flow_metadata_keys as metadata_keys
 import matching
@@ -30,49 +24,16 @@ from auth_base_forwarder import (
     resolved_auth_header_pairs,
 )
 from aws_sigv4 import AwsSigV4Credentials, AwsSigV4SigningError, sign_request
+from firewall_auth_cache import InvalidBillableAuthExpiryError, get_firewall_headers
+from firewall_auth_client import (
+    ConnectorNotConfiguredError,
+    FirewallAuthApiError,
+    InsufficientCreditsError,
+)
+from firewall_auth_config import auth_config_injects_credentials
 from logging_utils import log_proxy_entry
 from url_syntax import has_unsafe_runtime_url_syntax
 from url_utils import build_rewrite_url
-
-
-class ConnectorNotConfiguredError(Exception):
-    """Raised when the auth endpoint returns 424 — connector not linked or misconfigured."""
-
-
-class InsufficientCreditsError(Exception):
-    """Raised when the auth endpoint denies billable firewall auth for credits."""
-
-
-class InvalidBillableAuthExpiryError(Exception):
-    """Raised when billable firewall auth succeeds with an invalid cache expiry."""
-
-
-class FirewallAuthResponseTooLargeError(Exception):
-    """Raised when /firewall/auth returns a response body above the local cap."""
-
-
-class FirewallAuthApiError(Exception):
-    """Raised when /firewall/auth returns a structured error envelope."""
-
-    def __init__(
-        self,
-        *,
-        status: int,
-        code: str,
-        message: str,
-        connectors: list[str] | None = None,
-        failure_reason: str | None = None,
-    ):
-        super().__init__(message)
-        self.status = status
-        self.code = code
-        self.connectors = connectors
-        self.failure_reason = failure_reason
-
-
-class _ResponseBodyReader(Protocol):
-    def read(self, n: int = -1) -> bytes:
-        raise NotImplementedError
 
 
 class FirewallAuthHandlingResult(Enum):
@@ -83,58 +44,8 @@ class FirewallAuthHandlingResult(Enum):
     LOCAL_RESPONSE = "local_response"
 
 
-# Vercel bypass secret (still from environment as it's a secret)
-VERCEL_BYPASS = os.environ.get("VERCEL_AUTOMATION_BYPASS_SECRET", "")
-MAX_FIREWALL_AUTH_RESPONSE_BODY_BYTES = 256 * 1024
 _HTTP_STATUS_CLIENT_ERROR_MIN = 400
 _HTTP_STATUS_SERVER_ERROR_MIN = 500
-_STRUCTURED_FIREWALL_AUTH_ERROR_CODES = frozenset(
-    {
-        "FORBIDDEN",
-        "TOKEN_REFRESH_FAILED",
-        "TOKEN_ACCESS_RESOLUTION_FAILED",
-    }
-)
-_FIREWALL_AUTH_FAILURE_REASONS = frozenset({"upstream_provider", "reconnect_required"})
-
-
-@dataclass(frozen=True)
-class _FirewallAuthPayload:
-    """Cacheable /firewall/auth data applied to outbound requests."""
-
-    headers: dict[str, str]
-    resolved_secrets: list[str] = field(default_factory=list)
-    base: str | None = None
-    query: dict[str, str] | None = None
-    aws_sigv4: AwsSigV4Credentials | None = None
-
-
-@dataclass
-class _FirewallHeaderCacheEntry:
-    """Cached /firewall/auth response data for a single firewall key."""
-
-    payload: _FirewallAuthPayload
-    expires_at: object = None
-
-
-@dataclass(frozen=True)
-class _FirewallAuthSuccess:
-    """Validated /firewall/auth success response consumed by the auth cache."""
-
-    payload: _FirewallAuthPayload
-    expires_at: object = None
-    refreshed_connectors: list[str] = field(default_factory=list)
-    refreshed_secrets: list[str] = field(default_factory=list)
-
-
-@dataclass
-class _FirewallAuthState:
-    """Per-(run_id, api_id) auth lifecycle state."""
-
-    cache: _FirewallHeaderCacheEntry | None = None
-    lock: asyncio.Lock = field(default_factory=asyncio.Lock)
-    force_refresh_pending: bool = False
-    last_force_refresh_at: float = 0.0
 
 
 @dataclass(frozen=True)
@@ -156,42 +67,6 @@ class _FirewallAuthContext:
     secret_connector_metadata_map: dict | None
     vars_map: dict | None
     firewall_billable: bool
-
-
-_auth_state: dict[tuple[str, str], _FirewallAuthState] = {}
-
-# Cooldown window for re-marking a force-refresh. Caps amplification at
-# 30 refreshes/hour/key under a persistent non-token 401 loop — safely
-# below Google's 50/hour/user OAuth refresh limit (the tightest known).
-# The first force-refresh after a real token invalidation always fires
-# immediately; the cooldown only affects REPEATED forced refreshes, so
-# happy-path recovery is unaffected.
-_FORCE_REFRESH_COOLDOWN_SECS = 120.0
-
-
-def _non_empty_mapping(value: object) -> bool:
-    return isinstance(value, dict) and bool(value)
-
-
-def auth_config_injects_credentials(auth_config: object) -> bool:
-    """Return whether a firewall auth config can add managed credentials."""
-    if not isinstance(auth_config, dict):
-        return False
-    if _non_empty_mapping(auth_config.get("headers")):
-        return True
-    if _non_empty_mapping(auth_config.get("query")):
-        return True
-    if _non_empty_mapping(auth_config.get("awsSigv4")):
-        return True
-    return isinstance(auth_config.get("base"), str) and auth_config["base"] != ""
-
-
-def _get_auth_state(cache_key: tuple[str, str]) -> _FirewallAuthState:
-    state = _auth_state.get(cache_key)
-    if state is None:
-        state = _FirewallAuthState()
-        _auth_state[cache_key] = state
-    return state
 
 
 def is_billable_firewall(firewall_name: str, vm_info: dict) -> bool:
@@ -315,357 +190,6 @@ def _mark_matched_firewall_failure(
     flow.metadata[metadata_keys.FIREWALL_ERROR] = error_code
 
 
-def request_force_refresh(cache_key: tuple[str, str]) -> None:
-    """Request a forced token refresh on the next /firewall/auth fetch.
-
-    No-op if a forced refresh already completed within
-    ``_FORCE_REFRESH_COOLDOWN_SECS`` — rate limiter for the case where the
-    token is actually fine but the endpoint rejects for another reason
-    (scope, resource-level permission). See #9860.
-
-    Design notes for future changes:
-
-    * The consume timestamp in ``state.last_force_refresh_at`` is written **before**
-      ``fetch_firewall_headers`` is awaited in ``get_firewall_headers``, not
-      after. Recording after would allow a 401 arriving during the fetch to
-      re-add the marker; after the fetch completes and writes the cache, a
-      later cache miss would then consume the stale marker and trigger an
-      unnecessary second refresh. The trade-off is that a failed fetch
-      (webhook down, ``TOKEN_REFRESH_FAILED``, etc.) still burns the
-      cooldown — intentional, because if the refresh grant itself is broken,
-      retrying faster than once per cooldown wouldn't help and would hammer
-      the provider.
-    * ``time.time()`` is used for the cooldown (not ``time.monotonic()``) for
-      consistency with the rest of this module, which compares wall-clock
-      ``expiresAt`` values from the webhook. An NTP backward step could
-      freeze the cooldown until wall-clock catches up; on NTP-slewed runners
-      this is not a realistic concern.
-    """
-    state = _get_auth_state(cache_key)
-    if time.time() - state.last_force_refresh_at >= _FORCE_REFRESH_COOLDOWN_SECS:
-        state.force_refresh_pending = True
-
-
-def clear_cached_firewall_headers(cache_key: tuple[str, str]) -> None:
-    """Invalidate only cached headers while preserving refresh lifecycle state."""
-    state = _auth_state.get(cache_key)
-    if state:
-        state.cache = None
-
-
-def evict_stale_cache_keys(active_run_ids: set[str]) -> None:
-    """Remove cache entries for runs no longer in the registry."""
-    stale = [k for k in _auth_state if k[0] not in active_run_ids]
-    for k in stale:
-        _auth_state.pop(k, None)
-
-
-def evict_all_cache_keys() -> None:
-    """Remove all auth cache entries when active registry ownership is unknown."""
-    _auth_state.clear()
-
-
-def get_api_url() -> str:
-    """Get API URL from mitmproxy options."""
-    return ctx.options.vm0_api_url
-
-
-def make_api_request(url: str, data: bytes, sandbox_token: str) -> urllib.request.Request:
-    """Build a Request with standard platform API headers.
-
-    Centralises User-Agent, Authorization, Content-Type, and the optional
-    Vercel bypass header so that callers cannot accidentally omit them.
-    """
-    parsed_url = urllib.parse.urlsplit(url)
-    if parsed_url.scheme not in {"http", "https"} or not parsed_url.netloc:
-        raise ValueError("Platform API URL must be an absolute http(s) URL")
-
-    # S310 (suspicious-url-open-usage): callers build `url` from the
-    # operator-configured platform API URL, and the scheme is validated above
-    # before urllib can consume the request.
-    req = urllib.request.Request(  # noqa: S310
-        url,
-        data=data,
-        headers={
-            "Content-Type": "application/json",
-            "Authorization": f"Bearer {sandbox_token}",
-            "User-Agent": "vm0-mitm-addon/1.0",
-        },
-    )
-    if VERCEL_BYPASS:
-        req.add_header("x-vercel-protection-bypass", VERCEL_BYPASS)
-    return req
-
-
-def _read_firewall_auth_response_body(resp: _ResponseBodyReader) -> bytes:
-    body = resp.read(MAX_FIREWALL_AUTH_RESPONSE_BODY_BYTES + 1)
-    if len(body) > MAX_FIREWALL_AUTH_RESPONSE_BODY_BYTES:
-        raise FirewallAuthResponseTooLargeError("Firewall auth response body too large")
-    return body
-
-
-def _firewall_auth_api_error_from_envelope(
-    status: int,
-    error_info: dict,
-) -> FirewallAuthApiError | None:
-    code = error_info.get("code")
-    message = error_info.get("message")
-    if not isinstance(code, str) or not isinstance(message, str):
-        return None
-    if code not in _STRUCTURED_FIREWALL_AUTH_ERROR_CODES:
-        return None
-    connectors = error_info.get("connectors")
-    if isinstance(connectors, list) and all(isinstance(item, str) for item in connectors):
-        parsed_connectors = connectors
-    else:
-        parsed_connectors = None
-    failure_reason = error_info.get("failureReason")
-    parsed_failure_reason = (
-        failure_reason
-        if isinstance(failure_reason, str) and failure_reason in _FIREWALL_AUTH_FAILURE_REASONS
-        else None
-    )
-    return FirewallAuthApiError(
-        status=status,
-        code=code,
-        message=message,
-        connectors=parsed_connectors,
-        failure_reason=parsed_failure_reason,
-    )
-
-
-_MALFORMED_FIREWALL_AUTH_SUCCESS = "Firewall auth endpoint returned malformed success response"
-
-
-def _malformed_firewall_auth_success(message: str) -> ValueError:
-    return ValueError(f"{_MALFORMED_FIREWALL_AUTH_SUCCESS}: {message}")
-
-
-def _parse_string_map(value: object, field_name: str) -> dict[str, str]:
-    if not isinstance(value, dict):
-        raise _malformed_firewall_auth_success(f"{field_name} must be an object")
-
-    parsed: dict[str, str] = {}
-    for key, item in value.items():
-        if not isinstance(key, str):
-            raise _malformed_firewall_auth_success(f"{field_name} keys must be strings")
-        if not isinstance(item, str):
-            raise _malformed_firewall_auth_success(f"{field_name} values must be strings")
-        parsed[key] = item
-    return parsed
-
-
-def _parse_optional_string_map(
-    decoded: dict[object, object], field_name: str
-) -> dict[str, str] | None:
-    value = decoded.get(field_name)
-    if value is None:
-        return None
-    return _parse_string_map(value, field_name)
-
-
-def _parse_optional_string(decoded: dict[object, object], field_name: str) -> str | None:
-    value = decoded.get(field_name)
-    if value is None:
-        return None
-    if not isinstance(value, str):
-        raise _malformed_firewall_auth_success(f"{field_name} must be a string")
-    if value == "":
-        raise _malformed_firewall_auth_success(f"{field_name} must not be empty")
-    return value
-
-
-def _parse_optional_string_list(decoded: dict[object, object], field_name: str) -> list[str]:
-    value = decoded.get(field_name)
-    if value is None:
-        return []
-    if not isinstance(value, list):
-        raise _malformed_firewall_auth_success(f"{field_name} must be an array")
-    if not all(isinstance(item, str) for item in value):
-        raise _malformed_firewall_auth_success(f"{field_name} values must be strings")
-    return list(value)
-
-
-def _parse_optional_aws_sigv4_credentials(
-    decoded: dict[object, object],
-) -> AwsSigV4Credentials | None:
-    value = decoded.get("awsSigv4")
-    if value is None:
-        return None
-    if not isinstance(value, dict):
-        raise _malformed_firewall_auth_success("awsSigv4 must be an object")
-    access_key_id = value.get("accessKeyId")
-    secret_access_key = value.get("secretAccessKey")
-    if not isinstance(access_key_id, str) or not access_key_id:
-        raise _malformed_firewall_auth_success("awsSigv4.accessKeyId is required")
-    if not isinstance(secret_access_key, str) or not secret_access_key:
-        raise _malformed_firewall_auth_success("awsSigv4.secretAccessKey is required")
-    return AwsSigV4Credentials(
-        access_key_id=access_key_id,
-        secret_access_key=secret_access_key,
-        session_token=_parse_optional_string(value, "sessionToken"),
-    )
-
-
-def _parse_firewall_auth_success(decoded: object) -> _FirewallAuthSuccess:
-    if not isinstance(decoded, dict):
-        raise _malformed_firewall_auth_success("response must be an object")
-    decoded_map: dict[object, object] = decoded
-    if "headers" not in decoded_map:
-        raise _malformed_firewall_auth_success("headers is required")
-
-    base = decoded_map.get("base")
-    if base is not None and not isinstance(base, str):
-        raise _malformed_firewall_auth_success("base must be a string")
-
-    headers = _parse_string_map(decoded_map["headers"], "headers")
-    resolved_secrets = _parse_optional_string_list(decoded_map, "resolvedSecrets")
-    refreshed_connectors = _parse_optional_string_list(decoded_map, "refreshedConnectors")
-    refreshed_secrets = _parse_optional_string_list(decoded_map, "refreshedSecrets")
-    query = _parse_optional_string_map(decoded_map, "query")
-    aws_sigv4 = _parse_optional_aws_sigv4_credentials(decoded_map)
-    payload = _FirewallAuthPayload(
-        headers=headers,
-        resolved_secrets=resolved_secrets,
-        base=base,
-        query=query,
-        aws_sigv4=aws_sigv4,
-    )
-    return _FirewallAuthSuccess(
-        payload=payload,
-        expires_at=decoded_map.get("expiresAt"),
-        refreshed_connectors=refreshed_connectors,
-        refreshed_secrets=refreshed_secrets,
-    )
-
-
-def _fetch_firewall_headers_sync(
-    encrypted_secrets: str,
-    auth_headers: dict,
-    sandbox_token: str,
-    api_url: str,
-    *,
-    secret_connector_map: dict | None = None,
-    secret_connector_metadata_map: dict | None = None,
-    vars_map: dict | None = None,
-    auth_base: str | None = None,
-    auth_query: dict | None = None,
-    auth_aws_sigv4: dict | None = None,
-    firewall_billable: bool = False,
-    force_refresh: bool = False,
-) -> _FirewallAuthSuccess:
-    """Synchronous helper — runs in a thread to avoid blocking the event loop.
-
-    api_url is resolved by the async caller (fetch_firewall_headers) while
-    still on the event loop, so this function never touches ctx.options.
-    """
-    url = f"{api_url}/api/webhooks/agent/firewall/auth"
-    body: dict = {"encryptedSecrets": encrypted_secrets, "authHeaders": auth_headers}
-    if auth_base:
-        body["authBase"] = auth_base
-    if auth_query:
-        body["authQuery"] = auth_query
-    if auth_aws_sigv4:
-        body["authAwsSigv4"] = auth_aws_sigv4
-    if secret_connector_map:
-        body["secretConnectorMap"] = secret_connector_map
-    if secret_connector_metadata_map:
-        body["secretConnectorMetadataMap"] = secret_connector_metadata_map
-    if vars_map:
-        body["vars"] = vars_map
-    if firewall_billable:
-        body["firewallBillable"] = True
-    if force_refresh:
-        body["forceRefresh"] = True
-    data = json.dumps(body).encode()
-    req = make_api_request(url, data, sandbox_token)
-    try:
-        # nosemgrep: dynamic-urllib-use-detected
-        with urllib.request.urlopen(req, timeout=10) as resp:  # noqa: S310
-            decoded: object = json.loads(_read_firewall_auth_response_body(resp))
-            return _parse_firewall_auth_success(decoded)
-    except urllib.error.HTTPError as e:
-        # HTTPError wraps an open socket; `with e` closes on every exit
-        # path to avoid FD exhaustion under sustained cache-miss load (#10475).
-        with e:
-            try:
-                error_body = json.loads(_read_firewall_auth_response_body(e))
-            except (json.JSONDecodeError, OSError):
-                raise e from None
-            if not isinstance(error_body, dict):
-                raise e from None
-            error_info = error_body.get("error")
-            if not isinstance(error_info, dict):
-                raise e from None
-            error_message = error_info.get("message")
-            if error_info.get("code") == "CONNECTOR_NOT_CONFIGURED":
-                raise ConnectorNotConfiguredError(
-                    error_message if isinstance(error_message, str) else "Connector not configured",
-                ) from None
-            if error_info.get("code") == "INSUFFICIENT_CREDITS":
-                raise InsufficientCreditsError(
-                    error_message if isinstance(error_message, str) else "Insufficient credits",
-                ) from None
-            api_error = _firewall_auth_api_error_from_envelope(e.code, error_info)
-            if api_error is None:
-                raise e from None
-            raise api_error from None
-
-
-async def fetch_firewall_headers(
-    encrypted_secrets: str,
-    auth_headers: dict,
-    sandbox_token: str,
-    *,
-    secret_connector_map: dict | None = None,
-    secret_connector_metadata_map: dict | None = None,
-    vars_map: dict | None = None,
-    auth_base: str | None = None,
-    auth_query: dict | None = None,
-    auth_aws_sigv4: dict | None = None,
-    firewall_billable: bool = False,
-    force_refresh: bool = False,
-) -> _FirewallAuthSuccess:
-    """Resolve auth headers via server-side decryption.
-
-    encrypted_secrets is the encrypted runtime secret namespace. After API-side
-    decryption, keys are the `NAME` in `${{ secrets.NAME }}` and values are the
-    real secret values.
-
-    secret_connector_map maps firewall auth secret env aliases (the `NAME` in
-    `${{ secrets.NAME }}`) to the connector or provider owner that can
-    refresh/resolve access. secret_connector_metadata_map uses the same keys to
-    add source details when the owner alone is not enough to locate access
-    storage.
-
-    When secret_connector_map is provided, the auth endpoint can refresh
-    expired access tokens and returns an expiresAt timestamp for TTL caching.
-    For billable firewall auth, expiresAt is also bounded by the server-side
-    credit authorization lease.
-
-    When force_refresh is True, the endpoint refreshes access tokens regardless
-    of DB tokenExpiresAt — used after the upstream returns 401 (#9860).
-
-    Uses asyncio.to_thread to avoid blocking mitmproxy's event loop.
-    """
-    api_url = get_api_url()
-    return await asyncio.to_thread(
-        _fetch_firewall_headers_sync,
-        encrypted_secrets,
-        auth_headers,
-        sandbox_token,
-        api_url,
-        secret_connector_map=secret_connector_map,
-        secret_connector_metadata_map=secret_connector_metadata_map,
-        vars_map=vars_map,
-        auth_base=auth_base,
-        auth_query=auth_query,
-        auth_aws_sigv4=auth_aws_sigv4,
-        firewall_billable=firewall_billable,
-        force_refresh=force_refresh,
-    )
-
-
 def _merge_auth_headers(
     headers,
     auth_headers: dict[str, str],
@@ -682,143 +206,6 @@ def _merge_auth_headers(
     return [
         (name, value) for name, value in pairs if name.lower() not in override_names
     ] + auth_pairs
-
-
-def _has_valid_expiry(value: object, now: float | None = None) -> bool:
-    if isinstance(value, bool) or not isinstance(value, int | float):
-        return False
-    if not math.isfinite(value):
-        return False
-    return (time.time() if now is None else now) < value
-
-
-def _build_token_meta(
-    payload: _FirewallAuthPayload,
-    *,
-    cache_hit: bool,
-    refreshed_connectors: list[str] | None = None,
-    refreshed_secrets: list[str] | None = None,
-) -> dict:
-    token_meta: dict = {
-        "headers": payload.headers,
-        "resolved_secrets": payload.resolved_secrets,
-        "cache_hit": cache_hit,
-    }
-    if refreshed_connectors is not None:
-        token_meta["refreshed_connectors"] = refreshed_connectors
-    if refreshed_secrets is not None:
-        token_meta["refreshed_secrets"] = refreshed_secrets
-    if payload.base is not None:
-        token_meta["base"] = payload.base
-    if payload.query is not None:
-        token_meta["query"] = payload.query
-    if payload.aws_sigv4 is not None:
-        token_meta["aws_sigv4"] = payload.aws_sigv4
-    return token_meta
-
-
-def _build_cache_hit(
-    cached: _FirewallHeaderCacheEntry, firewall_billable: bool = False
-) -> dict | None:
-    """Check if a cached entry is still valid and return a cache-hit result."""
-    expires_at = cached.expires_at
-    now = time.time()
-    if expires_at is None:
-        if firewall_billable:
-            return None
-    elif not _has_valid_expiry(expires_at, now):
-        return None
-    return _build_token_meta(cached.payload, cache_hit=True)
-
-
-async def get_firewall_headers(
-    run_id: str,
-    api_id: str,
-    encrypted_secrets: str,
-    auth_headers: dict,
-    sandbox_token: str,
-    *,
-    secret_connector_map: dict | None = None,
-    secret_connector_metadata_map: dict | None = None,
-    vars_map: dict | None = None,
-    auth_base: str | None = None,
-    auth_query: dict | None = None,
-    auth_aws_sigv4: dict | None = None,
-    firewall_billable: bool = False,
-) -> dict:
-    """Get firewall auth headers with TTL-based caching.
-
-    Uses per-key locking so that concurrent requests for the same
-    (run_id, api_id) coalesce into a single HTTP fetch.
-
-    Cache is evicted when:
-    - The run is removed from the registry (see registry.load_registry)
-    - A 401 response is received (see response handler)
-    - The expiresAt timestamp from the auth endpoint has passed
-    """
-    cache_key = (run_id, api_id)
-    state = _get_auth_state(cache_key)
-
-    # Fast path: cache hit (no lock needed — single-threaded event loop)
-    if state.cache:
-        hit = _build_cache_hit(state.cache, firewall_billable=firewall_billable)
-        if hit:
-            return hit
-
-    # Slow path: acquire per-key lock so only one coroutine fetches
-    async with state.lock:
-        # Double-check: another coroutine may have populated cache while we waited
-        if state.cache:
-            hit = _build_cache_hit(state.cache, firewall_billable=firewall_billable)
-            if hit:
-                return hit
-
-        # Consume the force-refresh marker inside the lock so concurrent
-        # coroutines for the same (run_id, api_id) cannot both trigger a
-        # refresh — the one that loses the lock will see the fresh cache
-        # on its double-check above and never reach this path. Record the
-        # consume timestamp so request_force_refresh() suppresses re-marking
-        # within the cooldown window (guards against 401-amplification).
-        force_refresh = state.force_refresh_pending
-        state.force_refresh_pending = False
-        if force_refresh:
-            state.last_force_refresh_at = time.time()
-
-        result = await fetch_firewall_headers(
-            encrypted_secrets,
-            auth_headers,
-            sandbox_token,
-            secret_connector_map=secret_connector_map,
-            secret_connector_metadata_map=secret_connector_metadata_map,
-            vars_map=vars_map,
-            auth_base=auth_base,
-            auth_query=auth_query,
-            auth_aws_sigv4=auth_aws_sigv4,
-            firewall_billable=firewall_billable,
-            force_refresh=force_refresh,
-        )
-        if firewall_billable and not _has_valid_expiry(result.expires_at):
-            raise InvalidBillableAuthExpiryError(
-                "Billable firewall auth response did not include a valid cache expiry"
-            )
-        cache_entry = _FirewallHeaderCacheEntry(
-            payload=result.payload,
-            expires_at=result.expires_at,
-        )
-
-        # A 401 can request a forced refresh while this non-forced fetch is
-        # in flight. Return the current result to this request, but do not let
-        # it repopulate shared cache ahead of the pending forced refresh.
-        marker_appeared_during_non_forced_fetch = not force_refresh and state.force_refresh_pending
-        if not marker_appeared_during_non_forced_fetch:
-            state.cache = cache_entry
-
-        return _build_token_meta(
-            result.payload,
-            cache_hit=False,
-            refreshed_connectors=result.refreshed_connectors,
-            refreshed_secrets=result.refreshed_secrets,
-        )
 
 
 def _record_firewall_auth_success_metadata(flow: http.HTTPFlow, token_meta: dict) -> None:
