+++ crates/runner/mitm-addon/src/platform_api.py
@@ -1,0 +1,42 @@
+"""Shared platform API request helpers for the mitmproxy addon."""
+
+import os
+import urllib.parse
+import urllib.request
+
+from mitmproxy import ctx
+
+# Vercel bypass secret (still from environment as it's a secret)
+VERCEL_BYPASS = os.environ.get("VERCEL_AUTOMATION_BYPASS_SECRET", "")
+
+
+def get_api_url() -> str:
+    """Get API URL from mitmproxy options."""
+    return ctx.options.vm0_api_url
+
+
+def make_api_request(url: str, data: bytes, sandbox_token: str) -> urllib.request.Request:
+    """Build a Request with standard platform API headers.
+
+    Centralises User-Agent, Authorization, Content-Type, and the optional
+    Vercel bypass header so that callers cannot accidentally omit them.
+    """
+    parsed_url = urllib.parse.urlsplit(url)
+    if parsed_url.scheme not in {"http", "https"} or not parsed_url.netloc:
+        raise ValueError("Platform API URL must be an absolute http(s) URL")
+
+    # S310 (suspicious-url-open-usage): callers build `url` from the
+    # operator-configured platform API URL, and the scheme is validated above
+    # before urllib can consume the request.
+    req = urllib.request.Request(  # noqa: S310
+        url,
+        data=data,
+        headers={
+            "Content-Type": "application/json",
+            "Authorization": f"Bearer {sandbox_token}",
+            "User-Agent": "vm0-mitm-addon/1.0",
+        },
+    )
+    if VERCEL_BYPASS:
+        req.add_header("x-vercel-protection-bypass", VERCEL_BYPASS)
+    return req
